+++ src/zns.ts
@@ -92,4 +92,6 @@ export default class {
   isSupportedDomain(domain: string): boolean {
     return domain.indexOf('.') > 0 && /^.{1,}\.(zil)$/.test(domain);
   }
+
+  isSupportedNetwork(): boolean { return true }
 }